+++ projects/plugins/crm/src/automation/workflow/class-workflow-repository.php
@@ -3,7 +3,7 @@
  * The workflow repository responsible for communicating with the database.
  *
  * @package automattic/jetpack-crm
- * @since $$next-version$$
+ * @since 6.2.0
  */
 
 namespace Automattic\Jetpack\CRM\Automation\Workflow;
@@ -15,14 +15,14 @@ use wpdb;
 /**
  * Class Workflow_Repository.
  *
- * @since $$next-version$$
+ * @since 6.2.0
  */
 class Workflow_Repository {
 
 	/**
 	 * The WordPress database access layer.
 	 *
-	 * @since $$next-version$$
+	 * @since 6.2.0
 	 * @var wpdb
 	 */
 	protected $wpdb;
@@ -39,7 +39,7 @@ class Workflow_Repository {
 	 *
 	 * @global wpdb     $wpdb WordPress database abstraction object.
 	 * @global string[] $ZBSCRM_t An array of Jetpack CRM table names.
-	 * @since $$next-version$$
+	 * @since 6.2.0
 	 */
 	public function __construct() {
 		global $wpdb, $ZBSCRM_t; // phpcs:ignore WordPress.NamingConventions.ValidVariableName.VariableNotSnakeCase
@@ -51,7 +51,7 @@ class Workflow_Repository {
 	/**
 	 * Find Workflow by ID.
 	 *
-	 * @since $$next-version$$
+	 * @since 6.2.0
 	 *
 	 * @param int $id The workflow ID.
 	 * @return Automation_Workflow|false The workflow object or false if not found.
@@ -72,7 +72,7 @@ class Workflow_Repository {
 	/**
 	 * Get all workflows.
 	 *
-	 * @since $$next-version$$
+	 * @since 6.2.0
 	 *
 	 * @return Automation_Workflow[]
 	 */
@@ -85,7 +85,7 @@ class Workflow_Repository {
 	 *
 	 * @todo Implement "order by" logic.
 	 *
-	 * @since $$next-version$$
+	 * @since 6.2.0
 	 *
 	 * @param array  $criteria Workflow arguments to filter the workflows result.
 	 * @param string $order_by The column to order by.
@@ -154,7 +154,7 @@ class Workflow_Repository {
 	 *
 	 * This is used to both update and create a workflow.
 	 *
-	 * @since $$next-version$$
+	 * @since 6.2.0
 	 *
 	 * @param Automation_Workflow $workflow The workflow to persist.
 	 * @return void
@@ -172,7 +172,7 @@ class Workflow_Repository {
 	/**
 	 * Prepare the data to persist.
 	 *
-	 * @since $$next-version$$
+	 * @since 6.2.0
 	 *
 	 * @param Automation_Workflow $workflow The workflow to persist.
 	 * @return array The workflow raw data.
@@ -189,7 +189,7 @@ class Workflow_Repository {
 	/**
 	 * Insert a workflow into the database.
 	 *
-	 * @since $$next-version$$
+	 * @since 6.2.0
 	 *
 	 * @param Automation_Workflow $workflow The workflow to persist.
 	 *
@@ -226,7 +226,7 @@ class Workflow_Repository {
 	/**
 	 * Update a workflow.
 	 *
-	 * @since $$next-version$$
+	 * @since 6.2.0
 	 *
 	 * @param Automation_Workflow $workflow The workflow to persist.
 	 * @return void
@@ -255,7 +255,7 @@ class Workflow_Repository {
 	/**
 	 * Delete a workflow.
 	 *
-	 * @since $$next-version$$
+	 * @since 6.2.0
 	 *
 	 * @param Automation_Workflow $workflow The workflow to delete.
 	 * @return bool
@@ -286,7 +286,7 @@ class Workflow_Repository {
 	/**
 	 * Map a database row to a workflow object.
 	 *
-	 * @since $$next-version$$
+	 * @since 6.2.0
 	 *
 	 * @param array $row The database row.
 	 * @return Automation_Workflow